refactor(Distance): 거리를 포장한 Distance 추가
- meter <-> kilometer 단위 변환 처리

+++ src/main/resources/schema.sql
@@ -19,7 +19,7 @@ CREATE TABLE IF NOT EXISTS section
     line_id         BIGINT                NOT NULL,
     up_station_id   BIGINT                NOT NULL,
     down_station_id BIGINT                NOT NULL,
-    distance        INT,
+    distance        double,
     index_num       BIGINT,
     PRIMARY KEY (id),
     FOREIGN KEY (line_id) REFERENCES line (id),
